+++ lib/transpec/syntax/mixin/expect_base.rb
@@ -17,6 +17,10 @@ module Transpec
           add_matcher Receive
         end
 
+        def dynamic_analysis_target?
+          super && receiver_node.nil? && matcher_node
+        end
+
         def current_syntax_type
           :expect
         end
@@ -36,15 +40,21 @@ module Transpec
         end
 
         def to_node
+          return nil unless parent_node
+
           if parent_node.block_type? && parent_node.children.first.equal?(node)
             parent_node.parent_node
-          else
+          elsif parent_node.send_type? && parent_node.children.first.equal?(node)
             parent_node
+          else
+            nil
           end
         end
 
         def matcher_node
+          return nil unless to_node
           to_arg_node = to_node.children[2]
+          return nil unless to_arg_node
           Util.each_forward_chained_node(to_arg_node, :include_origin)
             .select(&:send_type?).to_a.last
         end